Provide custom fuzzer for BootstrapTokenString kubeadm type
This custom function is needed when ConvertTo/ConvertFrom functions
uses the json package to unmarshal the bootstrap token string.

The Kubeadm v1beta1.BootstrapTokenString type ships with a custom
json string representation, in particular it supplies a customized
UnmarshalJSON function that can return an error if the string
isn't in the correct form.

This function effectively disables any fuzzing for the token by setting
the values for ID and Secret to working alphanumeric values.

Signed-off-by: Vince Prignano <vincepri@vmware.com>